Gathering loader types and adding comments
* Gathering loader types regarding state, messages, action and types into 1 file
* Adding comment about loader screen listener and how it can be used
