+++ src/modules/loader/reducers.ts
@@ -1,17 +1,13 @@
-import {LoaderMsgs} from '~/types/loaderMessages';
-import {LoaderActions, LoaderTypes} from './types';
+import {LoaderStateI, LoaderMsgs} from '~/types/loader';
+
+import {LoaderActions, LoaderTypes} from '~/types/loader';
 
 type Actions = {
   type: LoaderActions;
   payload?: any;
 };
 
-interface StateI {
-  type: LoaderTypes;
-  msg: LoaderMsgs;
-}
-
-const initialState: StateI = {
+const initialState: LoaderStateI = {
   type: LoaderTypes.default,
   msg: LoaderMsgs.empty,
 };
@@ -20,7 +16,7 @@ const reducer = (state = initialState, action: Actions) => {
   switch (action.type) {
     case LoaderActions.set:
       return action.payload;
-    case LoaderActions.reset:
+    case LoaderActions.dismiss:
       return initialState;
     default:
       return state;
