+++ src/modules/loader/selectors.ts
@@ -1,6 +1,10 @@
 import {createSelector} from 'reselect';
 
-export const getLoaderState = (state) => state.loader;
+import {RootReducerI} from '~/types/reducer';
+import {LoaderStateI} from '~/types/loader';
+
+export const getLoaderState = (state: RootReducerI): LoaderStateI =>
+  state.loader;
 
 export const getLoaderType = createSelector(
   getLoaderState,
