--- src/modules/loader/types.ts
@@ -1,10 +1,0 @@
-export enum LoaderTypes {
-  default,
-  error,
-  success,
-}
-
-export enum LoaderActions {
-  reset,
-  set,
-}
