--- src/types/loaderMessages.ts
@@ -1,7 +1,0 @@
-export enum LoaderMsgs {
-  creating = 'Creating your personal secret number',
-  matching = 'Matching two instances',
-  success = 'Success!',
-  failed = 'Failed',
-  empty = '',
-}
