+++ src/types/reducer.ts
@@ -1,0 +1,5 @@
+import {LoaderStateI} from '~/types/loader';
+
+export interface RootReducerI {
+  loader: LoaderStateI;
+}
